ODK_Central/With_Pygsheets/Automatic ODK Submissions Tracker+Email Notification/requirements.txt
pandas
pygsheets
python-dotenv
schedule
pyodk
